mime: Convert built-in MIME mappings to PHP arrays
Currently, MimeAnalyzer builds the internal mappings of MIME types <=> file
extensions by concatenating several string buffers in mime.type format into a
giant string, and then parsing it. The mapping of MIME types to internal media
types is built up in a similar way, except we use a dubious homegrown format
with undocumented conventions. It's a mess, and an expensive one -- ~1.5% of
api.php CPU time on the WMF cluster is spent building these buffers
and parsing them. Converting the mappings to PHP associative arrays makes them
much cheaper to load and easier to maintain.

Doing this without breaking compatibility with existing behaviors requires some
delicate footwork. The current mime.types buffer is made up of the following
fragments, in order:

  1) MimeAnalyzer::$wellKnownTypes
  2) If $wgMimeTypeFile == 'includes/mime.types' (sic!):
       the contents of includes/libs/mime/mime.types.
     If $wgMimeTypeFile is another file path (e.g., '/etc/mime.types'):
       the contents of that file.
     If !wg$MimeTypeFile, this fragment is blank.
  3) MimeAnalyzer::$extraTypes (populated by extensions via hook).

The mime.info buffer is built up in the exact same way, except it's
MimeAnalyzer::$wellKnownInfo, $wgMimeInfoFile, and MimeAnalyzer::$extraInfo.

What this means in effect is that some built-in MediaWiki MIME mappings are
"baked in" (anything in MimeAnalyzer::$wellKnown*), and others can be
overridden (anything in includes/libs/mime/mime.*).

To avoid breaking backward compatibility, we have to preserve the distinction.
Thus this change has two MIME mappings, encapsulated in two classes:
'MimeMapMinimal', which contains just the baked-in mappings, and 'MimeMap'
which contains both the baked-in and overridable mappings.
We also have to keep the code for parsing mime.types and the ad-hoc mime.info
format, at least for now.

In a FUTURE change (i.e., not here), I think we can:

* Deprecate $wgMimeTypeFile in favor of a new config var, $wgExtraMimeTypeFile.
  $wgMimeTypeFile is evil because if you are using to add support for additional
  MIME types, you can end up unwittingly dropping support for other types that
  exist in MediaWiki's mime.types but not your file. The new $wgExtraMimeTypeFile
  would only be used to add new MIME mappings on top of the standard
  MimeMappings, which was probably the original intent for $wgMimeTypeFile.
* Deprecate $wgMimeInfoFile. I don't think we need to provide a replacement, because
  extensions can use the hook, and I doubt anyone is using the config var. But
  if we wanted to provide an alternative, we could have a $wgExtraMimeInfoMap
  that has an array of extra mappings.
* Deprecate MimeAnalyzer::addExtraTypes and MimeAnalyzer::addExtraInfo, and
  provide alternative interfaces that take structured input instead of string
  blobs.

Bug: T252228
Change-Id: I9b2979d3c9c0dee96bb19e0290f680724e718891